Remove git gate for dev tools install
Git comes pre-installed on some distros now, so it's
not a particularly good check for "are development
tools installed" or the real question, "are homebrew
dependencies installed". It might be a tad slower, but
the package managers can determine better than we
can if they need to do any work.